Revert TreatWChar_tAsBuiltInType=true setting, xerces from 3rd party is not built with the same setting and there are linking errors due to this.
Since UML2XML and Constraint Checker are COM components, it shouldn't
matter what xerces is GME using and how was it built so that UDM uses
the same xerces version in its 3rd party.

Due to lack of details, we presume that the probable cause of UML2XML
registration error is a bad path setting or wrong compilation.


git-svn-id: https://svn.isis.vanderbilt.edu/MoBIES/UDM/trunk@2877 c0adbae9-c137-0410-b742-ec4b2d2ceded