Add new Object and ObjectImpl methods to get/set the chain of references between a connection and its source/destination
The methods allow one to set the source and destination of a connection
together with a (possibly partial) chain of references to reach the source
or destination.

The methods are not implemented yet for DOM and static backends.


Make the new reference ports methods available in the generated API and test them

For now, Udm.exe will generate API methods for the new reference ports methods
if the name(s) of the association class are specified on the command line
with parameter --assoc-wrphelper.

git-svn-id: https://svn.isis.vanderbilt.edu/MoBIES/UDM/trunk@3468 c0adbae9-c137-0410-b742-ec4b2d2ceded